% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_call_stack.R
\name{complex_call_stack}
\alias{complex_call_stack}
\title{Example of a function that calls down into a deeply nested call stack}
\usage{
complex_call_stack(x)
}
\description{
Example of a function that calls down into a deeply nested call stack
}
