% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_call_stack.R
\name{deeper_nested_function}
\alias{deeper_nested_function}
\title{Example of a function within a deeply nested call stack}
\usage{
deeper_nested_function(x)
}
\description{
Example of a function within a deeply nested call stack
}
