% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{dependent_function}
\alias{dependent_function}
\title{Shows a dependency and pipe}
\usage{
dependent_function(x, n_cyl)
}
\arguments{
\item{x}{(actually mtcars dataset)}

\item{n_cyl}{number of cylinders}
}
\value{
returns a filtered mtcars
}
\description{
Shows a dependency and pipe
}
\examples{
dependent_function(x = mtcars, n_cyl = 4)
}
