% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{example}
\alias{example}
\title{Manually redefining a function from another package}
\usage{
example(
  topic,
  package = NULL,
  lib.loc = NULL,
  character.only = FALSE,
  give.lines = FALSE,
  local = FALSE,
  echo = TRUE,
  verbose = getOption("verbose"),
  setRNG = FALSE,
  ask = getOption("example.ask"),
  prompt.prefix = abbreviate(topic, 6),
  run.dontrun = FALSE,
  run.donttest = interactive()
)
}
\description{
Manually redefining a function from another package
}
