% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{give_me_y}
\alias{give_me_y}
\title{Function with an argument}
\usage{
give_me_y(x, n)
}
\arguments{
\item{x}{numeric}

\item{n}{numeric to multiply by}
}
\value{
returns  x * n
}
\description{
Function with an argument
}
\examples{
give_me_y(x = 3, n = 4)
}
