% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\name{hello2}
\alias{hello2}
\title{hello2}
\usage{
hello2(x = "Default Pasta")
}
\arguments{
\item{x}{Appended to the string that is printed out.}
}
\description{
Prins stuff.
}
\examples{
hello2()
hello2("Hello hello")
}
