% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypotenuse.R
\name{hypotenuse}
\alias{hypotenuse}
\title{Calculate the hypotenuse provided two edge lengths}
\usage{
hypotenuse(a, b)
}
\arguments{
\item{a, b}{edge lengths}
}
\description{
Calculate the hypotenuse provided two edge lengths
}
