% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_example.R
\name{names,S4Example-method}
\alias{names,S4Example-method}
\title{Provide names of the data slot in an S4Example object}
\usage{
\S4method{names}{S4Example}(x)
}
\arguments{
\item{x}{An x}
}
\description{
Provide names of the data slot in an S4Example object
}
