% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_example.R
\name{names,S4Example2-method}
\alias{names,S4Example2-method}
\title{Provide names of the data slot in an S4Example2 object, to ensure we can
distinguish methods}
\usage{
\S4method{names}{S4Example2}(x)
}
\arguments{
\item{x}{An x}
}
\description{
Provide names of the data slot in an S4Example2 object, to ensure we can
distinguish methods
}
