% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_call_stack.R
\name{nested_function}
\alias{nested_function}
\title{An example of a non-exported function in a deep call stack}
\usage{
nested_function(x)
}
\description{
An example of a non-exported function in a deep call stack
}
