% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_example.R
\name{Person}
\alias{Person}
\title{An example R6 Person class with public and private fields/methods}
\description{
An example R6 Person class with public and private fields/methods

An example R6 Person class with public and private fields/methods
}
\references{
\url{https://adv-r.hadley.nz/r6.html}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Person$new()}}
\item \href{#method-print}{\code{Person$print()}}
\item \href{#method-clone}{\code{Person$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Person
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$new(name, age = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A name for the person}

\item{\code{age}}{An age for the person}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Person objects info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments unused}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Person$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
