% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_example.R
\docType{data}
\name{PersonPrime}
\alias{PersonPrime}
\title{A materialized R6 object declared within a package namespace}
\format{
An object of class \code{Person} (inherits from \code{R6}) of length 4.
}
\usage{
PersonPrime
}
\description{
A materialized R6 object declared within a package namespace
}
\keyword{datasets}
