% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_affirmations_bar}
\alias{plot_affirmations_bar}
\title{Create a bar chart of our affirmations}
\usage{
plot_affirmations_bar(affirmations)
}
\arguments{
\item{affirmations}{A vector of affirmations from \code{\link[=affirmation]{affirmation()}}}
}
\description{
Create a bar chart of our affirmations
}
