% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Tell someone they did a good job}
\usage{
praise(name)
}
\arguments{
\item{name}{A string that is the name of a person}
}
\description{
Tell someone they did a good job
}
\examples{
praise("Babka")

}
