% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_data_sampler.R
\docType{data}
\name{rd_data_sampler}
\alias{rd_data_sampler}
\title{Rd data sampler}
\format{
A data.frame with 3 rows and 2 variables:
\describe{
\item{x}{Numeric values}
\item{y}{Character values}
}
}
\source{
<r.d.sampler/>
}
\usage{
rd_data_sampler
}
\description{
An example of dataset documentation
}
\examples{
rd_data_sampler

}
\keyword{datasets}
