% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_sampler.R
\name{rd_sampler}
\alias{rd_sampler}
\title{Rd sampler title}
\usage{
rd_sampler(x, y = TRUE, ...)
}
\arguments{
\item{x}{Rd sampler param}

\item{y}{Rd sampler param with default}

\item{...}{Rd sampler ellipsis param}
}
\value{
Rd sampler return
}
\description{
Rd sampler description with \href{www.example.com}{Rd sampler link}, \verb{Rd sampler in-line code}. And Rd dynamic content, \emph{italics text},
\emph{emphasis text}.
}
\details{
Rd sampler details

Rd sampler enumerated list
\enumerate{
\item One
\item Two
\item Three
}

Rd sampler itemized list
\itemize{
\item One
\item Two
\item Three
}\tabular{lll}{
   Rd \tab Sampler \tab Table \cr
   rd \tab sampler \tab table \cr
}


\eqn{Rd + sampler + inline + equation}

\deqn{Rd * sampler * block * equation}
}
\note{
Rd sampler note
}
\section{Rd sampler subsection}{
Rd sampler subsection text
\subsection{Rd sampler sub-subsection}{

Rd sampler sub-subsection text
}
}

\examples{
rd_sampler()

}
\references{
R.D. Sampler. (1999)
}
\seealso{
base::print
}
\author{
R.D. Sampler.
}
\concept{rd_samplers}
