% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_call_stack.R
\name{recursive_function}
\alias{recursive_function}
\title{An example of an internal recursive function}
\usage{
recursive_function(x, i = 3L)
}
\description{
An example of an internal recursive function
}
