% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{data}
\name{reexport_pkg_srcrefs}
\alias{reexport_pkg_srcrefs}
\title{Reexport with srcref}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
reexport_pkg_srcrefs
}
\description{
Reexport with srcref
}
\keyword{datasets}
