% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_example.R
\name{s3_example_func}
\alias{s3_example_func}
\alias{s3_example_func.default}
\alias{s3_example_func.list}
\title{An example of an S3 generic}
\usage{
s3_example_func(x, ...)

\method{s3_example_func}{default}(x, ...)

\method{s3_example_func}{list}(x, ...)
}
\arguments{
\item{x}{An X}

\item{...}{Ellipsis

Used to ensure traces can be linked through s3 dispatch}
}
\description{
An example of an S3 generic
}
