% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_example.R
\docType{class}
\name{S4Example2-class}
\alias{S4Example2-class}
\alias{S4Example2}
\title{Another example S4 class}
\description{
Another example S4 class
}
\section{Slots}{

\describe{
\item{\code{data}}{Some data}
}}

