% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_example.R
\name{show,S4Example-method}
\alias{show,S4Example-method}
\title{Provide show of the data slot in an S4Example object}
\usage{
\S4method{show}{S4Example}(object)
}
\arguments{
\item{x}{An x}
}
\description{
Provide show of the data slot in an S4Example object
}
