% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-functions.R
\name{summarize}
\alias{summarize}
\title{Summarize data}
\usage{
summarize(df, col, summary)
}
\arguments{
\item{df}{Data frame to summarize}

\item{col}{Column to summarize}

\item{summary}{Function to summarize with}
}
\description{
Summarize data
}
