% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_example.R
\name{Accumulator}
\alias{Accumulator}
\title{An example R6 Accumulator class}
\arguments{
\item{sum}{An initial sum to begin accumulating}

\item{x}{A value to add to the accumulating sum}
}
\value{
A new \code{Accumulator} object

The \code{Accumulator} object after accumulating the new value
}
\description{
Create a new Accumulator

Add to the accumulating sum
}
\section{Fields}{

\describe{
\item{\code{sum}}{An accumulated sum}
}}

\references{
\url{https://adv-r.hadley.nz/r6.html}
}
