% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{bigger_than_10}
\alias{bigger_than_10}
\title{Tests if a number if bigger than 10}
\usage{
bigger_than_10(n)
}
\arguments{
\item{n}{numeric}
}
\value{
logical
}
\description{
Tests if a number if bigger than 10
}
\examples{
bigger_than_10(n = 11)
}
