% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_example.R
\name{Person}
\alias{Person}
\title{An example R6 Person class with public and private fields/methods}
\arguments{
\item{name}{A name for the person}

\item{age}{An age for the person}

\item{...}{Arguments unused}
}
\description{
Create a new Person

Print a Person objects info
}
\references{
\url{https://adv-r.hadley.nz/r6.html}
}
