% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_example.R
\name{Rando}
\alias{Rando}
\title{An example R6 Rando class using an active field}
\description{
An example R6 Rando class using an active field
}
\section{Fields}{

\describe{
\item{\code{random}}{A random number, generated uniquely upon access}
}}

\references{
\url{https://adv-r.hadley.nz/r6.html}
}
